refactor: split comb node and effect modules
