+++ crates/celox/src/logic_tree/comb.rs
@@ -1,12 +1,12 @@
+mod effect;
+mod node;
 mod path;
 mod state;
 
 pub use path::{LogicPath, LogicPathTarget};
 pub use state::{BoundaryMap, SymbolicStore};
 
-use std::{collections::BTreeSet, fmt, hash::Hash};
-
-use serde::{Deserialize, Serialize};
+use std::{collections::BTreeSet, hash::Hash};
 
 use crate::ParserError;
 use crate::context_width::get_context_width;
@@ -28,943 +28,12 @@ use veryl_analyzer::ir::{
 };
 use veryl_analyzer::value::{Value, byte_value_to_string};
 
+use effect::{CombEffectCollector, collect_comb_effects_statements, subtract_written_sensitivity};
 use state::{FunctionControlState, FunctionLoopControlState, LoopControlState};
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, Serialize, Deserialize)]
-pub struct NodeId(pub usize);
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(bound(
-    serialize = "A: Serialize + std::hash::Hash + Eq + Clone",
-    deserialize = "A: Deserialize<'de> + std::hash::Hash + Eq + Clone"
-))]
-pub struct SLTNodeArena<A: Hash + Eq + Clone> {
-    pub nodes: Vec<SLTNode<A>>,
-    #[serde(skip, default = "crate::HashMap::default")]
-    pub cache: crate::HashMap<SLTNode<A>, NodeId>,
-}
-
-impl<A: PartialEq + Hash + Eq + Clone> PartialEq for SLTNodeArena<A> {
-    fn eq(&self, other: &Self) -> bool {
-        self.nodes == other.nodes
-    }
-}
-
-impl<A: Eq + Hash + Clone> Eq for SLTNodeArena<A> {}
-
-impl<A: Hash + Eq + Clone> SLTNodeArena<A> {
-    pub fn new() -> Self {
-        Self {
-            nodes: Vec::new(),
-            cache: crate::HashMap::default(),
-        }
-    }
-
-    pub fn alloc(&mut self, node: SLTNode<A>) -> NodeId
-    where
-        A: Hash + Eq + Clone,
-    {
-        if let Some(id) = self.cache.get(&node) {
-            return *id;
-        }
-        let id = NodeId(self.nodes.len());
-        self.cache.insert(node.clone(), id);
-        self.nodes.push(node);
-        id
-    }
-
-    pub fn get(&self, id: NodeId) -> &SLTNode<A> {
-        &self.nodes[id.0]
-    }
-
-    pub fn display(&self, id: NodeId) -> NodeDisplay<'_, A> {
-        NodeDisplay { arena: self, id }
-    }
-}
-
-pub struct NodeDisplay<'a, A: Hash + Eq + Clone> {
-    arena: &'a SLTNodeArena<A>,
-    id: NodeId,
-}
-
-impl<'a, A: Hash + Eq + Clone + std::fmt::Display> std::fmt::Display for NodeDisplay<'a, A> {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "n{}: ", self.id.0)?;
-        self.arena.get(self.id).fmt_expression(f, self.arena)
-    }
-}
-
-impl<A: Hash + Eq + Clone> SLTNode<A> {
-    pub fn fmt_expression(
-        &self,
-        f: &mut std::fmt::Formatter<'_>,
-        arena: &SLTNodeArena<A>,
-    ) -> std::fmt::Result
-    where
-        A: std::fmt::Display,
-    {
-        match self {
-            SLTNode::Input {
-                variable,
-                index,
-                access,
-                ..
-            } => {
-                write!(f, "{}", variable)?;
-                for idx in index {
-                    write!(f, "n{}", idx.node.0)?;
-                    write!(f, "[(idx)")?;
-                    arena.get(idx.node).fmt_expression(f, arena)?;
-                    if idx.stride > 1 {
-                        write!(f, " * {}", idx.stride)?;
-                    }
-                    write!(f, "]")?;
-                }
-                if index.is_empty() {
-                    write!(f, "{}", access)?;
-                } else {
-                    // For array access, the 'access' field represents the bit-slice within the element.
-                    // If it targets a multi-dimensional array, indices are processed recursively.
-                    if access.lsb != 0 || access.msb != 0 {
-                        // This depends on how SLTNode::Input is used for arrays.
-                        // If it's a bit-slice of an array element, we show it.
-                        // write!(f, "{}", access)?;
-                    }
-                }
-                Ok(())
-            }
-            SLTNode::Constant(val, _mask, _width, _signed) => {
-                write!(f, "{}", val)
-            }
-            SLTNode::Binary(lhs, op, rhs) => {
-                write!(f, "(")?;
-                write!(f, "n{}:", lhs.0)?;
-                arena.get(*lhs).fmt_expression(f, arena)?;
-                let op_str = match op {
-                    crate::ir::BinaryOp::Add => "+",
-                    crate::ir::BinaryOp::Sub => "-",
-                    crate::ir::BinaryOp::Mul => "*",
-                    crate::ir::BinaryOp::Div => "/",
-                    crate::ir::BinaryOp::Rem => "%",
-                    crate::ir::BinaryOp::And => "&",
-                    crate::ir::BinaryOp::Or => "|",
-                    crate::ir::BinaryOp::Xor => "^",
-                    crate::ir::BinaryOp::Shl => "<<",
-                    crate::ir::BinaryOp::Shr => ">>",
-                    crate::ir::BinaryOp::Sar => ">>>",
-                    crate::ir::BinaryOp::Eq => "==",
-                    crate::ir::BinaryOp::Ne => "!=",
-                    crate::ir::BinaryOp::LtU | crate::ir::BinaryOp::LtS => "<",
-                    crate::ir::BinaryOp::LeU | crate::ir::BinaryOp::LeS => "<=",
-                    crate::ir::BinaryOp::GtU | crate::ir::BinaryOp::GtS => ">",
-                    crate::ir::BinaryOp::GeU | crate::ir::BinaryOp::GeS => ">=",
-                    crate::ir::BinaryOp::LogicAnd => "&&",
-                    crate::ir::BinaryOp::LogicOr => "||",
-                    crate::ir::BinaryOp::EqWildcard => "==?",
-                    crate::ir::BinaryOp::NeWildcard => "!=?",
-                };
-                write!(f, " {} ", op_str)?;
-                write!(f, "n{}:", rhs.0)?;
-                arena.get(*rhs).fmt_expression(f, arena)?;
-                write!(f, ")")
-            }
-            SLTNode::Unary(op, inner) => {
-                let op_str = match op {
-                    crate::ir::UnaryOp::Ident => "",
-                    crate::ir::UnaryOp::Minus => "-",
-                    crate::ir::UnaryOp::BitNot => "~",
-                    crate::ir::UnaryOp::LogicNot => "!",
-                    crate::ir::UnaryOp::And => "&", // reduction
-                    crate::ir::UnaryOp::Or => "|",
-                    crate::ir::UnaryOp::Xor => "^",
-                };
-                write!(f, "{}(", op_str)?;
-                write!(f, "n{}:", inner.0)?;
-                arena.get(*inner).fmt_expression(f, arena)?;
-                write!(f, ")")
-            }
-            SLTNode::Mux {
-                cond,
-                then_expr,
-                else_expr,
-            } => {
-                write!(f, "(")?;
-                write!(f, "n{}:", cond.0)?;
-                arena.get(*cond).fmt_expression(f, arena)?;
-                write!(f, " ? ")?;
-                write!(f, "n{}:", then_expr.0)?;
-                arena.get(*then_expr).fmt_expression(f, arena)?;
-                write!(f, " : ")?;
-                write!(f, "n{}:", else_expr.0)?;
-                arena.get(*else_expr).fmt_expression(f, arena)?;
-                write!(f, ")")
-            }
-            SLTNode::ForFold {
-                loop_var,
-                start,
-                end,
-                inclusive,
-                step,
-                step_op,
-                reverse,
-                result,
-                initials,
-                updates,
-                ..
-            } => {
-                let fmt_bound =
-                    |f: &mut std::fmt::Formatter<'_>, bound: &SLTLoopBound| -> std::fmt::Result {
-                        match bound {
-                            SLTLoopBound::Const(v) => write!(f, "{v}"),
-                            SLTLoopBound::Expr(node) => {
-                                write!(f, "n{}:", node.0)?;
-                                arena.get(*node).fmt_expression(f, arena)
-                            }
-                        }
-                    };
-                write!(f, "for {loop_var} in ")?;
-                if *reverse {
-                    write!(f, "rev ")?;
-                }
-                fmt_bound(f, start)?;
-                if *inclusive {
-                    write!(f, "..=")?;
-                } else {
-                    write!(f, "..")?;
-                }
-                fmt_bound(f, end)?;
-                if *step != 1 || *step_op != SLTStepOp::Add {
-                    write!(f, " step ")?;
-                    match step_op {
-                        SLTStepOp::Add => write!(f, "+=")?,
-                        SLTStepOp::Mul => write!(f, "*=")?,
-                        SLTStepOp::Shl => write!(f, "<<=")?,
-                    }
-                    write!(f, " {step}")?;
-                }
-                write!(f, " => {result} init[")?;
-                for (i, init) in initials.iter().enumerate() {
-                    if i > 0 {
-                        write!(f, "; ")?;
-                    }
-                    write!(f, "{} = n{}:", init.target, init.expr.0)?;
-                    arena.get(init.expr).fmt_expression(f, arena)?;
-                }
-                write!(f, "] {{ ")?;
-                for (i, update) in updates.iter().enumerate() {
-                    if i > 0 {
-                        write!(f, "; ")?;
-                    }
-                    write!(f, "{} = n{}:", update.target, update.expr.0)?;
-                    arena.get(update.expr).fmt_expression(f, arena)?;
-                }
-                write!(f, " }}")
-            }
-            SLTNode::Concat(parts) => {
-                write!(f, "{{")?;
-                for (i, (part, w)) in parts.iter().enumerate() {
-                    if i > 0 {
-                        write!(f, ", ")?;
-                    }
-                    write!(f, "n{}@{w}:", part.0)?;
-                    arena.get(*part).fmt_expression(f, arena)?;
-                }
-                write!(f, "}}")
-            }
-            SLTNode::Slice { expr, access } => {
-                write!(f, "n{}:", expr.0)?;
-                arena.get(*expr).fmt_expression(f, arena)?;
-                write!(f, "{}", access)
-            }
-        }
-    }
-}
-
-impl<A: Hash + Eq + Clone> Default for SLTNodeArena<A> {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
-pub struct SLTIndex {
-    pub node: NodeId,
-    pub stride: usize,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
-pub enum SLTLoopBound {
-    Const(usize),
-    Expr(NodeId),
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
-pub enum SLTStepOp {
-    Add,
-    Mul,
-    Shl,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
-#[serde(bound(
-    serialize = "A: Serialize + std::hash::Hash + Eq + Clone",
-    deserialize = "A: Deserialize<'de> + std::hash::Hash + Eq + Clone"
-))]
-pub struct SLTForUpdate<A: Hash + Eq + Clone> {
-    pub target: VarAtomBase<A>,
-    pub expr: NodeId,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
-pub struct SLTForEffect {
-    pub site_id: u32,
-    pub guard: Option<NodeId>,
-    pub emit_on_true: bool,
-    pub args: Vec<NodeId>,
-    pub fatal_error_code: Option<i64>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
-#[serde(
-    into = "SLTNodeSerde<A>",
-    from = "SLTNodeSerde<A>",
-    bound(
-        serialize = "A: Serialize + Clone",
-        deserialize = "A: Deserialize<'de>"
-    )
-)]
-pub enum SLTNode<A: Hash + Eq + Clone> {
-    Input {
-        variable: A,
-        signed: bool,
-        index: Vec<SLTIndex>,
-        access: BitAccess,
-    },
-    Constant(BigUint, BigUint, usize, bool),
-    Binary(NodeId, BinaryOp, NodeId),
-    Unary(UnaryOp, NodeId),
-    Mux {
-        cond: NodeId,
-        then_expr: NodeId,
-        else_expr: NodeId,
-    },
-    ForFold {
-        loop_var: A,
-        loop_width: usize,
-        loop_signed: bool,
-        start: SLTLoopBound,
-        end: SLTLoopBound,
-        inclusive: bool,
-        step: usize,
-        step_op: SLTStepOp,
-        reverse: bool,
-        result: VarAtomBase<A>,
-        initials: Vec<SLTForUpdate<A>>,
-        updates: Vec<SLTForUpdate<A>>,
-        effects: Vec<SLTForEffect>,
-        continue_cond: NodeId,
-    },
-    // Concat/Slice are primarily used for RHS expression evaluation.
-    // On the LHS (assignments), bit manipulation is handled implicitly by RangeStore atomization.
-    Concat(Vec<(NodeId, usize)>),
-    Slice {
-        expr: NodeId,
-        access: BitAccess,
-    },
-}
-
-/// Serde-friendly mirror of [`SLTNode`] that replaces [`BigUint`] with `Vec<u8>` (little-endian).
-#[derive(Serialize, Deserialize)]
-#[serde(bound(serialize = "A: Serialize", deserialize = "A: Deserialize<'de>"))]
-enum SLTNodeSerde<A: Hash + Eq + Clone> {
-    Input {
-        variable: A,
-        signed: bool,
-        index: Vec<SLTIndex>,
-        access: BitAccess,
-    },
-    Constant {
-        payload: Vec<u8>,
-        mask: Vec<u8>,
-        width: usize,
-        signed: bool,
-    },
-    Binary(NodeId, BinaryOp, NodeId),
-    Unary(UnaryOp, NodeId),
-    Mux {
-        cond: NodeId,
-        then_expr: NodeId,
-        else_expr: NodeId,
-    },
-    ForFold {
-        loop_var: A,
-        loop_width: usize,
-        loop_signed: bool,
-        start: SLTLoopBound,
-        end: SLTLoopBound,
-        inclusive: bool,
-        step: usize,
-        step_op: SLTStepOp,
-        reverse: bool,
-        result: VarAtomBase<A>,
-        initials: Vec<SLTForUpdate<A>>,
-        updates: Vec<SLTForUpdate<A>>,
-        effects: Vec<SLTForEffect>,
-        continue_cond: NodeId,
-    },
-    Concat(Vec<(NodeId, usize)>),
-    Slice {
-        expr: NodeId,
-        access: BitAccess,
-    },
-}
-
-impl<A: Hash + Eq + Clone> From<SLTNode<A>> for SLTNodeSerde<A> {
-    fn from(node: SLTNode<A>) -> Self {
-        match node {
-            SLTNode::Input {
-                variable,
-                signed,
-                index,
-                access,
-            } => SLTNodeSerde::Input {
-                variable,
-                signed,
-                index,
-                access,
-            },
-            SLTNode::Constant(payload, mask, width, signed) => SLTNodeSerde::Constant {
-                payload: payload.to_bytes_le(),
-                mask: mask.to_bytes_le(),
-                width,
-                signed,
-            },
-            SLTNode::Binary(a, op, b) => SLTNodeSerde::Binary(a, op, b),
-            SLTNode::Unary(op, a) => SLTNodeSerde::Unary(op, a),
-            SLTNode::Mux {
-                cond,
-                then_expr,
-                else_expr,
-            } => SLTNodeSerde::Mux {
-                cond,
-                then_expr,
-                else_expr,
-            },
-            SLTNode::ForFold {
-                loop_var,
-                loop_width,
-                loop_signed,
-                start,
-                end,
-                inclusive,
-                step,
-                step_op,
-                reverse,
-                result,
-                initials,
-                updates,
-                effects,
-                continue_cond,
-            } => SLTNodeSerde::ForFold {
-                loop_var,
-                loop_width,
-                loop_signed,
-                start,
-                end,
-                inclusive,
-                step,
-                step_op,
-                reverse,
-                result,
-                initials,
-                updates,
-                effects,
-                continue_cond,
-            },
-            SLTNode::Concat(parts) => SLTNodeSerde::Concat(parts),
-            SLTNode::Slice { expr, access } => SLTNodeSerde::Slice { expr, access },
-        }
-    }
-}
-
-impl<A: Hash + Eq + Clone> From<SLTNodeSerde<A>> for SLTNode<A> {
-    fn from(node: SLTNodeSerde<A>) -> Self {
-        match node {
-            SLTNodeSerde::Input {
-                variable,
-                signed,
-                index,
-                access,
-            } => SLTNode::Input {
-                variable,
-                signed,
-                index,
-                access,
-            },
-            SLTNodeSerde::Constant {
-                payload,
-                mask,
-                width,
-                signed,
-            } => SLTNode::Constant(
-                BigUint::from_bytes_le(&payload),
-                BigUint::from_bytes_le(&mask),
-                width,
-                signed,
-            ),
-            SLTNodeSerde::Binary(a, op, b) => SLTNode::Binary(a, op, b),
-            SLTNodeSerde::Unary(op, a) => SLTNode::Unary(op, a),
-            SLTNodeSerde::Mux {
-                cond,
-                then_expr,
-                else_expr,
-            } => SLTNode::Mux {
-                cond,
-                then_expr,
-                else_expr,
-            },
-            SLTNodeSerde::ForFold {
-                loop_var,
-                loop_width,
-                loop_signed,
-                start,
-                end,
-                inclusive,
-                step,
-                step_op,
-                reverse,
-                result,
-                initials,
-                updates,
-                effects,
-                continue_cond,
-            } => SLTNode::ForFold {
-                loop_var,
-                loop_width,
-                loop_signed,
-                start,
-                end,
-                inclusive,
-                step,
-                step_op,
-                reverse,
-                result,
-                initials,
-                updates,
-                effects,
-                continue_cond,
-            },
-            SLTNodeSerde::Concat(parts) => SLTNode::Concat(parts),
-            SLTNodeSerde::Slice { expr, access } => SLTNode::Slice { expr, access },
-        }
-    }
-}
-impl<A: fmt::Debug + fmt::Display + Hash + Eq + Clone> SLTNode<A> {
-    /// Maps the address type A to B recursively throughout the tree.
-    pub fn map_addr<B, F>(
-        &self,
-        id: NodeId,
-        arena: &SLTNodeArena<A>,
-        target_arena: &mut SLTNodeArena<B>,
-        cache: &mut HashMap<NodeId, NodeId>,
-        f: &F,
-    ) -> NodeId
-    where
-        A: Hash + Eq + Clone,
-        B: Hash + Eq + Clone,
-        F: Fn(&A) -> B,
-    {
-        if let Some(mapped_id) = cache.get(&id) {
-            return *mapped_id;
-        }
-
-        let new_node = match self {
-            // Leaf: Transform address A to B
-            SLTNode::Input {
-                variable: addr,
-                signed,
-                index,
-                access,
-            } => {
-                let mapped_index = index
-                    .iter()
-                    .map(|idx| SLTIndex {
-                        node: arena
-                            .get(idx.node)
-                            .map_addr(idx.node, arena, target_arena, cache, f),
-                        stride: idx.stride,
-                    })
-                    .collect();
-                SLTNode::Input {
-                    variable: f(addr),
-                    signed: *signed,
-                    index: mapped_index,
-                    access: *access,
-                }
-            }
-
-            // Leaf: Constants remain unchanged
-            SLTNode::Constant(val, mask, width, signed) => {
-                SLTNode::Constant(val.clone(), mask.clone(), *width, *signed)
-            }
-
-            // Recursive cases
-            SLTNode::Binary(lhs, op, rhs) => {
-                let l = arena
-                    .get(*lhs)
-                    .map_addr(*lhs, arena, target_arena, cache, f);
-                let r = arena
-                    .get(*rhs)
-                    .map_addr(*rhs, arena, target_arena, cache, f);
-                SLTNode::Binary(l, *op, r)
-            }
-
-            SLTNode::Unary(op, inner) => {
-                let i = arena
-                    .get(*inner)
-                    .map_addr(*inner, arena, target_arena, cache, f);
-                SLTNode::Unary(*op, i)
-            }
-
-            SLTNode::Mux {
-                cond,
-                then_expr,
-                else_expr,
-            } => {
-                let c = arena
-                    .get(*cond)
-                    .map_addr(*cond, arena, target_arena, cache, f);
-                let t = arena
-                    .get(*then_expr)
-                    .map_addr(*then_expr, arena, target_arena, cache, f);
-                let e = arena
-                    .get(*else_expr)
-                    .map_addr(*else_expr, arena, target_arena, cache, f);
-                SLTNode::Mux {
-                    cond: c,
-                    then_expr: t,
-                    else_expr: e,
-                }
-            }
-
-            SLTNode::ForFold {
-                loop_var,
-                loop_width,
-                loop_signed,
-                start,
-                end,
-                inclusive,
-                step,
-                step_op,
-                reverse,
-                result,
-                initials,
-                updates,
-                effects,
-                continue_cond,
-            } => {
-                let map_bound =
-                    |bound: &SLTLoopBound,
-                     cache: &mut HashMap<NodeId, NodeId>,
-                     target_arena: &mut SLTNodeArena<B>|
-                     -> SLTLoopBound {
-                        match bound {
-                            SLTLoopBound::Const(v) => SLTLoopBound::Const(*v),
-                            SLTLoopBound::Expr(node) => SLTLoopBound::Expr(
-                                arena
-                                    .get(*node)
-                                    .map_addr(*node, arena, target_arena, cache, f),
-                            ),
-                        }
-                    };
-                let mapped_initials = initials
-                    .iter()
-                    .map(|update| SLTForUpdate {
-                        target: VarAtomBase::new(
-                            f(&update.target.id),
-                            update.target.access.lsb,
-                            update.target.access.msb,
-                        ),
-                        expr: arena.get(update.expr).map_addr(
-                            update.expr,
-                            arena,
-                            target_arena,
-                            cache,
-                            f,
-                        ),
-                    })
-                    .collect();
-                let mapped_updates = updates
-                    .iter()
-                    .map(|update| SLTForUpdate {
-                        target: VarAtomBase::new(
-                            f(&update.target.id),
-                            update.target.access.lsb,
-                            update.target.access.msb,
-                        ),
-                        expr: arena.get(update.expr).map_addr(
-                            update.expr,
-                            arena,
-                            target_arena,
-                            cache,
-                            f,
-                        ),
-                    })
-                    .collect();
-                let mapped_effects = effects
-                    .iter()
-                    .map(|effect| SLTForEffect {
-                        site_id: effect.site_id,
-                        guard: effect.guard.map(|guard| {
-                            arena
-                                .get(guard)
-                                .map_addr(guard, arena, target_arena, cache, f)
-                        }),
-                        emit_on_true: effect.emit_on_true,
-                        args: effect
-                            .args
-                            .iter()
-                            .map(|arg| {
-                                arena
-                                    .get(*arg)
-                                    .map_addr(*arg, arena, target_arena, cache, f)
-                            })
-                            .collect(),
-                        fatal_error_code: effect.fatal_error_code,
-                    })
-                    .collect();
-                SLTNode::ForFold {
-                    loop_var: f(loop_var),
-                    loop_width: *loop_width,
-                    loop_signed: *loop_signed,
-                    start: map_bound(start, cache, target_arena),
-                    end: map_bound(end, cache, target_arena),
-                    inclusive: *inclusive,
-                    step: *step,
-                    step_op: *step_op,
-                    reverse: *reverse,
-                    result: VarAtomBase::new(f(&result.id), result.access.lsb, result.access.msb),
-                    initials: mapped_initials,
-                    updates: mapped_updates,
-                    effects: mapped_effects,
-                    continue_cond: arena.get(*continue_cond).map_addr(
-                        *continue_cond,
-                        arena,
-                        target_arena,
-                        cache,
-                        f,
-                    ),
-                }
-            }
-
-            SLTNode::Concat(parts) => {
-                let mapped_parts = parts
-                    .iter()
-                    .map(|(node, width)| {
-                        (
-                            arena
-                                .get(*node)
-                                .map_addr(*node, arena, target_arena, cache, f),
-                            *width,
-                        )
-                    })
-                    .collect();
-                SLTNode::Concat(mapped_parts)
-            }
-
-            SLTNode::Slice { expr, access } => {
-                let e = arena
-                    .get(*expr)
-                    .map_addr(*expr, arena, target_arena, cache, f);
-                SLTNode::Slice {
-                    expr: e,
-                    access: *access,
-                }
-            }
-        };
-        let new_id = target_arena.alloc(new_node);
-        cache.insert(id, new_id);
-        new_id
-    }
-}
-
-/// Display implementation for SLTNode - provides human-readable tree structure
-///
-/// This implementation formats the Signal Logic Tree (SLT) as a hierarchical ASCII tree,
-/// making it easy to visualize the expression structure. Each node type displays relevant
-/// information:
-///
-/// - **Input**: Shows variable ID, dynamic indices (if any), and bit range [lsb:msb]
-/// - **Constant**: Displays the value in hexadecimal and width in bits
-/// - **Binary**: Shows the operation and recursively formats both operands with indentation
-/// - **Unary**: Shows the operation and recursively formats the inner expression
-/// - **Mux**: Displays condition, then-branch, and else-branch with clear labels
-/// - **Concat**: Lists concatenated parts with their widths
-/// - **Slice**: Shows the bit range extraction with the inner expression
-///
-/// # Example
-///
-/// A binary expression `a + b` would display as:
-/// ```text
-/// Binary(Add)
-///   Const(0x1, 32bits)
-///   Const(0x2, 32bits)
-/// ```
-///
-/// A more complex expression `(a + b) * (c - d)` would show:
-/// ```text
-/// Binary(Mul)
-///   Binary(Add)
-///     Const(0x1, 32bits)
-///     Const(0x2, 32bits)
-///   Binary(Sub)
-///     Const(0x3, 32bits)
-///     Const(0x4, 32bits)
-/// ```
-impl<A: fmt::Debug + fmt::Display + Hash + Eq + Clone> SLTNode<A> {
-    pub fn fmt_display(&self, f: &mut fmt::Formatter<'_>, arena: &SLTNodeArena<A>) -> fmt::Result {
-        self.fmt_recursive(f, 0, arena)
-    }
-}
-
-impl<A: fmt::Debug + fmt::Display + Hash + Eq + Clone> SLTNode<A> {
-    fn fmt_recursive(
-        &self,
-        f: &mut fmt::Formatter<'_>,
-        depth: usize,
-        arena: &SLTNodeArena<A>,
-    ) -> fmt::Result {
-        let indent = "  ".repeat(depth);
-        let child_indent = "  ".repeat(depth + 1);
-        match self {
-            SLTNode::Input {
-                variable,
-                index,
-                access,
-                ..
-            } => {
-                write!(f, "{}Input({:?}", indent, variable)?;
-                if !index.is_empty() {
-                    write!(f, "[")?;
-                    for (i, idx) in index.iter().enumerate() {
-                        if i > 0 {
-                            write!(f, ", ")?;
-                        }
-                        write!(f, "n{}:...", idx.node.0)?;
-                        if idx.stride > 1 {
-                            write!(f, "*{}", idx.stride)?;
-                        }
-                    }
-                    write!(f, "]")?;
-                }
-                write!(f, "[{}:{}]", access.lsb, access.msb)?;
-                write!(f, ")")
-            }
-            SLTNode::Constant(val, _mask, width, _signed) => {
-                write!(f, "{}Const({:#x}, {}bits)", indent, val, width)
-            }
-            SLTNode::Binary(lhs, op, rhs) => {
-                let op_str = format!("{:?}", op); // Just use Debug for simplicity
-                writeln!(f, "{}Binary({})", indent, op_str)?;
-                arena.get(*lhs).fmt_recursive(f, depth + 1, arena)?;
-                writeln!(f)?; // Insert empty line between left and right expressions
-                arena.get(*rhs).fmt_recursive(f, depth + 1, arena)
-            }
-            SLTNode::Unary(op, inner) => {
-                writeln!(f, "{}Unary({:?})", indent, op)?;
-                arena.get(*inner).fmt_recursive(f, depth + 1, arena)
-            }
-            SLTNode::Mux {
-                cond,
-                then_expr,
-                else_expr,
-            } => {
-                writeln!(f, "{}Mux", indent)?;
-                writeln!(f, "{}cond:", child_indent)?;
-                arena.get(*cond).fmt_recursive(f, depth + 2, arena)?;
-                writeln!(f, "\n{}then:", child_indent)?;
-                arena.get(*then_expr).fmt_recursive(f, depth + 2, arena)?;
-                writeln!(f, "\n{}else:", child_indent)?;
-                arena.get(*else_expr).fmt_recursive(f, depth + 2, arena)
-            }
-            SLTNode::ForFold {
-                loop_var,
-                loop_width,
-                loop_signed,
-                start,
-                end,
-                inclusive,
-                step,
-                step_op,
-                reverse,
-                result,
-                initials,
-                updates,
-                effects,
-                continue_cond,
-            } => {
-                writeln!(
-                    f,
-                    "{}ForFold(loop_var={}, width={}, signed={}, inclusive={}, step={}, step_op={:?}, reverse={}, result={})",
-                    indent,
-                    loop_var,
-                    loop_width,
-                    loop_signed,
-                    inclusive,
-                    step,
-                    step_op,
-                    reverse,
-                    result
-                )?;
-                writeln!(f, "{}start: {:?}", child_indent, start)?;
-                writeln!(f, "{}end: {:?}", child_indent, end)?;
-                for init in initials {
-                    writeln!(f, "{}init {}:", child_indent, init.target)?;
-                    arena.get(init.expr).fmt_recursive(f, depth + 2, arena)?;
-                    writeln!(f)?;
-                }
-                for update in updates {
-                    writeln!(f, "{}update {}:", child_indent, update.target)?;
-                    arena.get(update.expr).fmt_recursive(f, depth + 2, arena)?;
-                    writeln!(f)?;
-                }
-                for effect in effects {
-                    writeln!(f, "{}effect site={}:", child_indent, effect.site_id)?;
-                    if let Some(guard) = effect.guard {
-                        writeln!(f, "{}guard:", child_indent)?;
-                        arena.get(guard).fmt_recursive(f, depth + 2, arena)?;
-                        writeln!(f)?;
-                    }
-                    for arg in &effect.args {
-                        writeln!(f, "{}arg:", child_indent)?;
-                        arena.get(*arg).fmt_recursive(f, depth + 2, arena)?;
-                        writeln!(f)?;
-                    }
-                }
-                writeln!(f, "{}continue:", child_indent)?;
-                arena
-                    .get(*continue_cond)
-                    .fmt_recursive(f, depth + 2, arena)?;
-                Ok(())
-            }
-            SLTNode::Concat(parts) => {
-                writeln!(f, "{}Concat", indent)?;
-                for (i, (part, width)) in parts.iter().enumerate() {
-                    if i > 0 {
-                        writeln!(f)?;
-                    }
-                    writeln!(f, "{}[{}bits]:", child_indent, width)?;
-                    arena.get(*part).fmt_recursive(f, depth + 2, arena)?;
-                }
-                Ok(())
-            }
-            SLTNode::Slice { expr, access } => {
-                writeln!(f, "{}Slice[{}:{}]", indent, access.lsb, access.msb)?;
-                arena.get(*expr).fmt_recursive(f, depth + 1, arena)
-            }
-        }
-    }
-}
+pub use node::{
+    NodeId, SLTForEffect, SLTForUpdate, SLTIndex, SLTLoopBound, SLTNode, SLTNodeArena, SLTStepOp,
+};
 
 pub fn parse_comb(
     module: &Module,
@@ -1082,1034 +151,6 @@ pub fn parse_comb(
         effects.observers,
         effects.sites,
     ))
-}
-
-fn subtract_written_sensitivity<A: Copy + Eq + std::hash::Hash>(
-    atoms: impl IntoIterator<Item = VarAtomBase<A>>,
-    written_atoms: &[VarAtomBase<A>],
-) -> HashSet<VarAtomBase<A>> {
-    let mut result = HashSet::default();
-    for atom in atoms {
-        let mut ranges = vec![(atom.access.lsb, atom.access.msb)];
-        for written in written_atoms {
-            if written.id != atom.id {
-                continue;
-            }
-            ranges = ranges
-                .into_iter()
-                .flat_map(|(lsb, msb)| {
-                    if written.access.msb < lsb || written.access.lsb > msb {
-                        return vec![(lsb, msb)];
-                    }
-                    let mut kept = Vec::new();
-                    if lsb < written.access.lsb {
-                        kept.push((lsb, written.access.lsb - 1));
-                    }
-                    if written.access.msb < msb {
-                        kept.push((written.access.msb + 1, msb));
-                    }
-                    kept
-                })
-                .collect();
-        }
-        for (lsb, msb) in ranges {
-            result.insert(VarAtomBase::new(atom.id, lsb, msb));
-        }
-    }
-    result
-}
-
-#[derive(Default)]
-struct CombEffectCollector {
-    observers: Vec<CombObserver<VarId>>,
-    sites: Vec<RuntimeEventSite>,
-    sensitivity: HashSet<VarAtomBase<VarId>>,
-    active_guard: Option<NodeId>,
-    active_guard_sources: HashSet<VarAtomBase<VarId>>,
-    loop_effects: Option<Vec<SLTForEffect>>,
-}
-
-fn static_string_expr(expr: &Expression) -> Option<String> {
-    if !expr.comptime().r#type.is_string() {
-        return None;
-    }
-    let value = expr.comptime().get_value().ok()?;
-    byte_value_to_string(value)
-}
-
-fn register_comb_runtime_event_site<'a>(
-    collector: &mut CombEffectCollector,
-    kind: RuntimeEventKind,
-    args: &'a [SystemFunctionInput],
-) -> (u32, &'a [SystemFunctionInput]) {
-    let (template, value_args) = if args
-        .first()
-        .and_then(|arg| static_string_expr(&arg.0))
-        .is_some()
-    {
-        (
-            args.first().and_then(|arg| static_string_expr(&arg.0)),
-            &args[1..],
-        )
-    } else {
-        (None, args)
-    };
-    let site = RuntimeEventSite {
-        kind,
-        template,
-        arg_widths: value_args
-            .iter()
-            .map(|arg| arg.0.comptime().r#type.total_width().unwrap_or(1))
-            .collect(),
-        arg_signed: value_args
-            .iter()
-            .map(|arg| arg.0.comptime().expr_context.signed)
-            .collect(),
-        arg_is_string: value_args
-            .iter()
-            .map(|arg| arg.0.comptime().r#type.is_string())
-            .collect(),
-    };
-    let id = collector.sites.len() as u32;
-    collector.sites.push(site);
-    (id, value_args)
-}
-
-fn collect_system_function_effect(
-    module: &Module,
-    store: &SymbolicStore<VarId>,
-    call: &SystemFunctionCall,
-    arena: &mut SLTNodeArena<VarId>,
-    collector: &mut CombEffectCollector,
-) -> Result<(), ParserError> {
-    let (kind, cond, args) = match &call.kind {
-        SystemFunctionKind::Display(args) | SystemFunctionKind::Write(args) => {
-            (RuntimeEventKind::Display, None, args.as_slice())
-        }
-        SystemFunctionKind::Assert { kind, cond, args } => {
-            let event_kind = match kind {
-                veryl_analyzer::ir::AssertKind::Fatal => RuntimeEventKind::AssertFatal,
-                veryl_analyzer::ir::AssertKind::Continue => RuntimeEventKind::AssertContinue,
-            };
-            (event_kind, Some(&cond.0), args.as_slice())
-        }
-        _ => {
-            return Err(ParserError::unsupported(
-                66,
-                LoweringPhase::CombLowering,
-                "system function call",
-                format!("{call}"),
-                Some(&call.comptime.token),
-            ));
-        }
-    };
-    let (site_id, value_args) = register_comb_runtime_event_site(collector, kind, args);
-    let mut observer_args = Vec::new();
-    let mut observed_inputs = HashSet::default();
-    let mut position_inputs = HashSet::default();
-    for arg in value_args {
-        collect_expression_effects(module, store, &arg.0, arena, collector)?;
-        let ((node, sources), _) = eval_expression(module, store, &arg.0, arena, None)?;
-        observed_inputs.extend(sources.iter().copied());
-        collector.sensitivity.extend(sources);
-        collect_expression_position_inputs(module, &arg.0, &mut position_inputs)?;
-        observer_args.push(node);
-    }
-    let explicit_guard = if let Some(cond) = cond {
-        collect_expression_effects(module, store, cond, arena, collector)?;
-        let ((cond_node, cond_sources), _) = eval_expression(module, store, cond, arena, None)?;
-        observed_inputs.extend(cond_sources.iter().copied());
-        collector.sensitivity.extend(cond_sources);
-        collect_expression_position_inputs(module, cond, &mut position_inputs)?;
-        Some(cond_node)
-    } else {
-        None
-    };
-    observed_inputs.extend(collector.active_guard_sources.iter().copied());
-    let guard = match (kind, collector.active_guard, explicit_guard) {
-        (RuntimeEventKind::Display, active, None) => active,
-        (RuntimeEventKind::AssertContinue | RuntimeEventKind::AssertFatal, None, explicit) => {
-            explicit
-        }
-        (
-            RuntimeEventKind::AssertContinue | RuntimeEventKind::AssertFatal,
-            Some(active),
-            Some(explicit),
-        ) => {
-            let inactive = arena.alloc(SLTNode::Unary(UnaryOp::LogicNot, active));
-            Some(arena.alloc(SLTNode::Binary(inactive, BinaryOp::LogicOr, explicit)))
-        }
-        (RuntimeEventKind::AssertContinue | RuntimeEventKind::AssertFatal, Some(active), None) => {
-            Some(active)
-        }
-        (RuntimeEventKind::Display, _, Some(_)) => unreachable!("display has no explicit guard"),
-    };
-    let loop_effect = collector.loop_effects.as_ref().map(|_| SLTForEffect {
-        site_id,
-        guard,
-        emit_on_true: matches!(kind, RuntimeEventKind::Display),
-        args: observer_args.clone(),
-        fatal_error_code: matches!(kind, RuntimeEventKind::AssertFatal)
-            .then_some(1_000_000 + site_id as i64),
-    });
-    let observed_ids: HashSet<_> = observed_inputs.iter().map(|atom| atom.id).collect();
-    let position_ids: HashSet<_> = position_inputs.iter().map(|atom| atom.id).collect();
-    let preceding_writes: Vec<_> = store
-        .iter()
-        .flat_map(|(id, range_store)| {
-            range_store
-                .ranges
-                .iter()
-                .filter_map(move |(&lsb, (value, width, _))| {
-                    value
-                        .is_some()
-                        .then_some(VarAtomBase::new(*id, lsb, lsb + width - 1))
-                })
-        })
-        .collect();
-    collector.observers.push(CombObserver {
-        site_id,
-        activation_group: 0,
-        guard,
-        args: observer_args,
-        loop_runner: None,
-        sensitivity: Vec::new(),
-        local_inputs: store
-            .iter()
-            .filter_map(|(id, range_store)| {
-                if !observed_ids.contains(id) {
-                    return None;
-                }
-                if guard.is_none() {
-                    let overlaps_observed_input =
-                        range_store.ranges.iter().any(|(&lsb, (value, width, _))| {
-                            value.is_some()
-                                && observed_inputs.iter().chain(position_inputs.iter()).any(
-                                    |atom| {
-                                        atom.id == *id
-                                            && atom
-                                                .access
-                                                .overlaps(&BitAccess::new(lsb, lsb + width - 1))
-                                    },
-                                )
-                        });
-                    if !overlaps_observed_input {
-                        return None;
-                    }
-                }
-                let width = module
-                    .variables
-                    .get(id)
-                    .and_then(|var| resolve_total_width(module, var).ok())?;
-                if width == 0 {
-                    return None;
-                }
-                let parts = range_store.get_parts(BitAccess::new(0, width - 1));
-                let modified = parts.iter().any(|(value, _)| value.is_some());
-                if !modified {
-                    return None;
-                }
-                let (node, _) = combine_parts_with_default(*id, 0, parts, arena);
-                Some((*id, node))
-            })
-            .collect(),
-        observed_inputs: observed_inputs.into_iter().collect(),
-        position_inputs: position_inputs.into_iter().collect(),
-        preceding_writes: preceding_writes.clone(),
-        written_before: store
-            .iter()
-            .filter(|(id, _)| position_ids.contains(id))
-            .flat_map(|(id, range_store)| {
-                range_store
-                    .ranges
-                    .iter()
-                    .filter_map(move |(&lsb, (value, width, _))| {
-                        value
-                            .is_some()
-                            .then_some(VarAtomBase::new(*id, lsb, lsb + width - 1))
-                    })
-            })
-            .collect(),
-        written_input_atoms: Vec::new(),
-        written_inputs: Vec::new(),
-        captured_in_loop: loop_effect.is_some(),
-    });
-    if let (Some(loop_effects), Some(loop_effect)) = (&mut collector.loop_effects, loop_effect) {
-        loop_effects.push(loop_effect);
-    }
-    Ok(())
-}
-
-fn with_collector_guard<T, F>(
-    collector: &mut CombEffectCollector,
-    arena: &mut SLTNodeArena<VarId>,
-    guard: NodeId,
-    guard_sources: HashSet<VarAtomBase<VarId>>,
-    f: F,
-) -> Result<T, ParserError>
-where
-    F: FnOnce(&mut CombEffectCollector, &mut SLTNodeArena<VarId>) -> Result<T, ParserError>,
-{
-    let saved_guard = collector.active_guard;
-    let saved_guard_sources = collector.active_guard_sources.clone();
-    let active_guard = if let Some(active) = saved_guard {
-        arena.alloc(SLTNode::Binary(active, BinaryOp::LogicAnd, guard))
-    } else {
-        guard
-    };
-    let mut active_guard_sources = saved_guard_sources.clone();
-    active_guard_sources.extend(guard_sources);
-    collector.active_guard = Some(active_guard);
-    collector.active_guard_sources = active_guard_sources;
-    let result = f(collector, arena);
-    collector.active_guard = saved_guard;
-    collector.active_guard_sources = saved_guard_sources;
-    result
-}
-
-fn collect_function_call_effects(
-    module: &Module,
-    store: &SymbolicStore<VarId>,
-    call: &veryl_analyzer::ir::FunctionCall,
-    arena: &mut SLTNodeArena<VarId>,
-    collector: &mut CombEffectCollector,
-) -> Result<(), ParserError> {
-    let Some(function) = module.functions.get(&call.id) else {
-        return Err(ParserError::unsupported(
-            62,
-            LoweringPhase::CombLowering,
-            "function call",
-            format!("unknown function id: {:?}", call.id),
-            Some(&call.comptime.token),
-        ));
-    };
-
-    let Some(function_body) = (if let Some(index) = &call.index {
-        function.get_function(index)
-    } else {
-        function.get_function(&[])
-    }) else {
-        return Err(ParserError::unsupported(
-            62,
-            LoweringPhase::CombLowering,
-            "function call specialization",
-            format!("{call}"),
-            Some(&call.comptime.token),
-        ));
-    };
-
-    let mut local_store = store.clone();
-    for (arg_path, arg_id) in &function_body.arg_map {
-        let Some(arg_expr) = call.inputs.get(arg_path) else {
-            continue;
-        };
-        collect_expression_effects(module, store, arg_expr, arena, collector)?;
-
-        let formal = &module.variables[arg_id];
-        let arg_width = resolve_total_width(module, formal)?;
-        let ((arg_node, arg_sources), _) =
-            eval_expression(module, store, arg_expr, arena, Some(arg_width))?;
-        local_store.insert(
-            *arg_id,
-            RangeStore::new(Some((arg_node, arg_sources)), arg_width),
-        );
-    }
-
-    if !statements_contain_runtime_effect(module, &function_body.statements) {
-        return Ok(());
-    }
-
-    if let Some(ret_id) = function_body.ret {
-        collect_function_body_effects(
-            module,
-            local_store,
-            &function_body.statements,
-            ret_id,
-            arena,
-            collector,
-        )?;
-    } else {
-        let _ = collect_comb_effects_statements(
-            module,
-            local_store,
-            &function_body.statements,
-            arena,
-            collector,
-        )?;
-    }
-    Ok(())
-}
-
-fn statements_contain_runtime_effect(module: &Module, statements: &[Statement]) -> bool {
-    statements
-        .iter()
-        .any(|stmt| statement_contains_runtime_effect(module, stmt))
-}
-
-fn statement_contains_runtime_effect(module: &Module, stmt: &Statement) -> bool {
-    match stmt {
-        Statement::SystemFunctionCall(call) => matches!(
-            call.kind,
-            SystemFunctionKind::Display(_)
-                | SystemFunctionKind::Write(_)
-                | SystemFunctionKind::Assert { .. }
-        ),
-        Statement::If(if_stmt) => {
-            statements_contain_runtime_effect(module, &if_stmt.true_side)
-                || statements_contain_runtime_effect(module, &if_stmt.false_side)
-        }
-        Statement::For(for_stmt) => statements_contain_runtime_effect(module, &for_stmt.body),
-        Statement::FunctionCall(call) => module
-            .functions
-            .get(&call.id)
-            .and_then(|function| {
-                if let Some(index) = &call.index {
-                    function.get_function(index)
-                } else {
-                    function.get_function(&[])
-                }
-            })
-            .is_some_and(|body| statements_contain_runtime_effect(module, &body.statements)),
-        Statement::Assign(_)
-        | Statement::IfReset(_)
-        | Statement::TbMethodCall(_)
-        | Statement::Break
-        | Statement::Unsupported(_)
-        | Statement::Null => false,
-    }
-}
-
-fn collect_expression_effects(
-    module: &Module,
-    store: &SymbolicStore<VarId>,
-    expr: &Expression,
-    arena: &mut SLTNodeArena<VarId>,
-    collector: &mut CombEffectCollector,
-) -> Result<(), ParserError> {
-    match expr {
-        Expression::Term(factor) => collect_factor_effects(module, store, factor, arena, collector),
-        Expression::Unary(_, inner, _) => {
-            collect_expression_effects(module, store, inner, arena, collector)
-        }
-        Expression::Binary(lhs, _, rhs, _) => {
-            collect_expression_effects(module, store, lhs, arena, collector)?;
-            collect_expression_effects(module, store, rhs, arena, collector)
-        }
-        Expression::Ternary(cond, then_expr, else_expr, _) => {
-            collect_expression_effects(module, store, cond, arena, collector)?;
-            collect_expression_effects(module, store, then_expr, arena, collector)?;
-            collect_expression_effects(module, store, else_expr, arena, collector)
-        }
-        Expression::Concatenation(items, _) => {
-            for (item_expr, repeat_expr) in items {
-                collect_expression_effects(module, store, item_expr, arena, collector)?;
-                if let Some(repeat_expr) = repeat_expr {
-                    collect_expression_effects(module, store, repeat_expr, arena, collector)?;
-                }
-            }
-            Ok(())
-        }
-        Expression::ArrayLiteral(items, _) => {
-            for item in items {
-                match item {
-                    ArrayLiteralItem::Value(item_expr, repeat_expr) => {
-                        collect_expression_effects(module, store, item_expr, arena, collector)?;
-                        if let Some(repeat_expr) = repeat_expr {
-                            collect_expression_effects(
-                                module,
-                                store,
-                                repeat_expr,
-                                arena,
-                                collector,
-                            )?;
-                        }
-                    }
-                    ArrayLiteralItem::Defaul(default_expr) => {
-                        collect_expression_effects(module, store, default_expr, arena, collector)?;
-                    }
-                }
-            }
-            Ok(())
-        }
-        Expression::StructConstructor(_, fields, _) => {
-            for (_, field_expr) in fields {
-                collect_expression_effects(module, store, field_expr, arena, collector)?;
-            }
-            Ok(())
-        }
-    }
-}
-
-fn collect_factor_effects(
-    module: &Module,
-    store: &SymbolicStore<VarId>,
-    factor: &Factor,
-    arena: &mut SLTNodeArena<VarId>,
-    collector: &mut CombEffectCollector,
-) -> Result<(), ParserError> {
-    match factor {
-        Factor::Variable(_, index, select, _) => {
-            for expr in index.0.iter().chain(select.0.iter()) {
-                collect_expression_effects(module, store, expr, arena, collector)?;
-            }
-            if let Some((_, expr)) = &select.1 {
-                collect_expression_effects(module, store, expr, arena, collector)?;
-            }
-            Ok(())
-        }
-        Factor::FunctionCall(call) => {
-            collect_function_call_effects(module, store, call, arena, collector)
-        }
-        Factor::SystemFunctionCall(call) => match &call.kind {
-            SystemFunctionKind::Bits(input)
-            | SystemFunctionKind::Size(input)
-            | SystemFunctionKind::Clog2(input)
-            | SystemFunctionKind::Onehot(input)
-            | SystemFunctionKind::Signed(input)
-            | SystemFunctionKind::Unsigned(input) => {
-                collect_expression_effects(module, store, &input.0, arena, collector)
-            }
-            _ => Ok(()),
-        },
-        Factor::Value(_) | Factor::Anonymous(_) | Factor::Unknown(_) => Ok(()),
-    }
-}
-
-fn collect_function_body_effects(
-    module: &Module,
-    local_store: SymbolicStore<VarId>,
-    statements: &[Statement],
-    ret_id: VarId,
-    arena: &mut SLTNodeArena<VarId>,
-    collector: &mut CombEffectCollector,
-) -> Result<(), ParserError> {
-    fn collect_statements(
-        module: &Module,
-        mut state: FunctionControlState,
-        statements: &[Statement],
-        ret_id: VarId,
-        arena: &mut SLTNodeArena<VarId>,
-        collector: &mut CombEffectCollector,
-    ) -> Result<FunctionControlState, ParserError> {
-        for stmt in statements {
-            if matches!(constant_bool(arena, state.live_expr), Some(false)) {
-                break;
-            }
-            state = collect_statement(module, state, stmt, ret_id, arena, collector)?;
-        }
-        Ok(state)
-    }
-
-    fn collect_statement(
-        module: &Module,
-        state: FunctionControlState,
-        stmt: &Statement,
-        ret_id: VarId,
-        arena: &mut SLTNodeArena<VarId>,
-        collector: &mut CombEffectCollector,
-    ) -> Result<FunctionControlState, ParserError> {
-        if matches!(constant_bool(arena, state.live_expr), Some(false)) {
-            return Ok(state);
-        }
-        match stmt {
-            Statement::Assign(assign) => {
-                let store = state.store.clone();
-                let live = state.live_expr;
-                let live_sources = state.live_sources.clone();
-                with_collector_guard(collector, arena, live, live_sources, |collector, arena| {
-                    collect_expression_effects(module, &store, &assign.expr, arena, collector)
-                })?;
-                let (store, boundaries) =
-                    eval_assign(module, state.store, state.boundaries, assign, arena)?;
-                let live_expr = if function_assigns_whole_var(assign, ret_id) {
-                    bool_node(arena, false)
-                } else {
-                    bool_node(arena, true)
-                };
-                Ok(FunctionControlState {
-                    store,
-                    boundaries,
-                    live_expr,
-                    live_sources: HashSet::default(),
-                })
-            }
-            Statement::SystemFunctionCall(call) => {
-                let store = state.store.clone();
-                let live = state.live_expr;
-                let live_sources = state.live_sources.clone();
-                with_collector_guard(collector, arena, live, live_sources, |collector, arena| {
-                    collect_system_function_effect(module, &store, call, arena, collector)
-                })?;
-                Ok(state)
-            }
-            Statement::FunctionCall(call) => {
-                let store = state.store.clone();
-                let live = state.live_expr;
-                let live_sources = state.live_sources.clone();
-                with_collector_guard(collector, arena, live, live_sources, |collector, arena| {
-                    collect_function_call_effects(module, &store, call, arena, collector)
-                })?;
-                let (store, boundaries) = eval_statement_form_function_call(
-                    module,
-                    state.store,
-                    state.boundaries,
-                    call,
-                    arena,
-                    LoweringPhase::CombLowering,
-                )?;
-                Ok(FunctionControlState {
-                    store,
-                    boundaries,
-                    live_expr: bool_node(arena, true),
-                    live_sources: HashSet::default(),
-                })
-            }
-            Statement::If(if_stmt) => {
-                let store = state.store.clone();
-                let live = state.live_expr;
-                let live_sources = state.live_sources.clone();
-                with_collector_guard(collector, arena, live, live_sources, |collector, arena| {
-                    collect_expression_effects(module, &store, &if_stmt.cond, arena, collector)
-                })?;
-
-                let ((cond_node, cond_sources), _) =
-                    eval_expression(module, &state.store, &if_stmt.cond, arena, None)?;
-                let true_guard = arena.alloc(SLTNode::Binary(
-                    state.live_expr,
-                    BinaryOp::LogicAnd,
-                    cond_node,
-                ));
-                let mut true_sources = state.live_sources.clone();
-                true_sources.extend(cond_sources.iter().copied());
-                with_collector_guard(
-                    collector,
-                    arena,
-                    true_guard,
-                    true_sources,
-                    |collector, arena| {
-                        let _ = collect_statements(
-                            module,
-                            state.clone(),
-                            &if_stmt.true_side,
-                            ret_id,
-                            arena,
-                            collector,
-                        )?;
-                        Ok(())
-                    },
-                )?;
-
-                let false_cond = arena.alloc(SLTNode::Unary(UnaryOp::LogicNot, cond_node));
-                let false_guard = arena.alloc(SLTNode::Binary(
-                    state.live_expr,
-                    BinaryOp::LogicAnd,
-                    false_cond,
-                ));
-                let mut false_sources = state.live_sources.clone();
-                false_sources.extend(cond_sources.iter().copied());
-                with_collector_guard(
-                    collector,
-                    arena,
-                    false_guard,
-                    false_sources,
-                    |collector, arena| {
-                        let _ = collect_statements(
-                            module,
-                            state.clone(),
-                            &if_stmt.false_side,
-                            ret_id,
-                            arena,
-                            collector,
-                        )?;
-                        Ok(())
-                    },
-                )?;
-
-                let (store, boundaries) =
-                    eval_if(module, state.store, state.boundaries, if_stmt, arena)?;
-                Ok(FunctionControlState {
-                    store,
-                    boundaries,
-                    live_expr: bool_node(arena, true),
-                    live_sources: HashSet::default(),
-                })
-            }
-            Statement::For(for_stmt) => {
-                let store = state.store.clone();
-                let live = state.live_expr;
-                let live_sources = state.live_sources.clone();
-                with_collector_guard(collector, arena, live, live_sources, |collector, arena| {
-                    let _ = collect_comb_effects_for(module, store, for_stmt, arena, collector)?;
-                    Ok(())
-                })?;
-                let (store, boundaries) =
-                    eval_for(module, state.store, state.boundaries, for_stmt, arena)?;
-                Ok(FunctionControlState {
-                    store,
-                    boundaries,
-                    live_expr: bool_node(arena, true),
-                    live_sources: HashSet::default(),
-                })
-            }
-            Statement::Null => Ok(state),
-            Statement::IfReset(ir) => Err(ParserError::illegal_context(
-                "statement in comb function body",
-                format!("{stmt}"),
-                Some(&ir.token),
-            )),
-            Statement::TbMethodCall(_) | Statement::Break | Statement::Unsupported(_) => {
-                Err(ParserError::illegal_context(
-                    "statement in comb function body",
-                    format!("{stmt}"),
-                    None,
-                ))
-            }
-        }
-    }
-
-    let _ = collect_statements(
-        module,
-        FunctionControlState {
-            store: local_store,
-            boundaries: BoundaryMap::default(),
-            live_expr: bool_node(arena, true),
-            live_sources: HashSet::default(),
-        },
-        statements,
-        ret_id,
-        arena,
-        collector,
-    )?;
-    Ok(())
-}
-
-fn collect_expression_position_inputs(
-    module: &Module,
-    expr: &Expression,
-    out: &mut HashSet<VarAtomBase<VarId>>,
-) -> Result<(), ParserError> {
-    match expr {
-        Expression::Term(factor) => collect_factor_position_inputs(module, factor, out),
-        Expression::Binary(lhs, _, rhs, _) => {
-            collect_expression_position_inputs(module, lhs, out)?;
-            collect_expression_position_inputs(module, rhs, out)
-        }
-        Expression::Unary(_, inner, _) => collect_expression_position_inputs(module, inner, out),
-        Expression::Ternary(cond, then_expr, else_expr, _) => {
-            collect_expression_position_inputs(module, cond, out)?;
-            collect_expression_position_inputs(module, then_expr, out)?;
-            collect_expression_position_inputs(module, else_expr, out)
-        }
-        Expression::Concatenation(parts, _) => {
-            for (part, repeat) in parts {
-                collect_expression_position_inputs(module, part, out)?;
-                if let Some(repeat) = repeat {
-                    collect_expression_position_inputs(module, repeat, out)?;
-                }
-            }
-            Ok(())
-        }
-        Expression::ArrayLiteral(items, _) => {
-            for item in items {
-                match item {
-                    ArrayLiteralItem::Value(expr, repeat) => {
-                        collect_expression_position_inputs(module, expr, out)?;
-                        if let Some(repeat) = repeat {
-                            collect_expression_position_inputs(module, repeat, out)?;
-                        }
-                    }
-                    ArrayLiteralItem::Defaul(expr) => {
-                        collect_expression_position_inputs(module, expr, out)?;
-                    }
-                }
-            }
-            Ok(())
-        }
-        Expression::StructConstructor(_, fields, _) => {
-            for (_, field_expr) in fields {
-                collect_expression_position_inputs(module, field_expr, out)?;
-            }
-            Ok(())
-        }
-    }
-}
-
-fn collect_factor_position_inputs(
-    module: &Module,
-    factor: &Factor,
-    out: &mut HashSet<VarAtomBase<VarId>>,
-) -> Result<(), ParserError> {
-    match factor {
-        Factor::Variable(var_id, index, select, comptime) => {
-            if !comptime.is_const {
-                let access = eval_var_select(module, *var_id, index, select)?;
-                out.insert(VarAtomBase::new(*var_id, access.lsb, access.msb));
-            }
-            for expr in index.0.iter().chain(select.0.iter()) {
-                collect_expression_position_inputs(module, expr, out)?;
-            }
-            if let Some((_, expr)) = &select.1 {
-                collect_expression_position_inputs(module, expr, out)?;
-            }
-            Ok(())
-        }
-        Factor::Value(_) => Ok(()),
-        Factor::FunctionCall(call) => {
-            for arg in call.inputs.values() {
-                collect_expression_position_inputs(module, arg, out)?;
-            }
-            Ok(())
-        }
-        Factor::SystemFunctionCall(_) | Factor::Anonymous(_) | Factor::Unknown(_) => Ok(()),
-    }
-}
-
-fn collect_comb_effects_statements(
-    module: &Module,
-    mut store: SymbolicStore<VarId>,
-    statements: &[Statement],
-    arena: &mut SLTNodeArena<VarId>,
-    collector: &mut CombEffectCollector,
-) -> Result<SymbolicStore<VarId>, ParserError> {
-    for stmt in statements {
-        match stmt {
-            Statement::Assign(assign) => {
-                collect_expression_effects(module, &store, &assign.expr, arena, collector)?;
-                let (next_store, _) =
-                    eval_assign(module, store, BoundaryMap::default(), assign, arena)?;
-                store = next_store;
-            }
-            Statement::SystemFunctionCall(call) => {
-                collect_system_function_effect(module, &store, call, arena, collector)?;
-            }
-            Statement::FunctionCall(call) => {
-                collect_function_call_effects(module, &store, call, arena, collector)?;
-                let (next_store, _) = eval_statement_form_function_call(
-                    module,
-                    store,
-                    BoundaryMap::default(),
-                    call,
-                    arena,
-                    LoweringPhase::CombLowering,
-                )?;
-                store = next_store;
-            }
-            Statement::For(for_stmt) => {
-                store = collect_comb_effects_for(module, store, for_stmt, arena, collector)?;
-            }
-            Statement::If(if_stmt) => {
-                collect_expression_effects(module, &store, &if_stmt.cond, arena, collector)?;
-                let ((cond_node, sources), _) =
-                    eval_expression(module, &store, &if_stmt.cond, arena, None)?;
-                collector.sensitivity.extend(sources.iter().copied());
-                let saved_guard = collector.active_guard;
-                let saved_guard_sources = collector.active_guard_sources.clone();
-                let true_guard = if let Some(active) = saved_guard {
-                    arena.alloc(SLTNode::Binary(active, BinaryOp::LogicAnd, cond_node))
-                } else {
-                    cond_node
-                };
-                let mut true_sources = saved_guard_sources.clone();
-                true_sources.extend(sources.iter().copied());
-                collector.active_guard = Some(true_guard);
-                collector.active_guard_sources = true_sources;
-                let side_store = collect_comb_effects_statements(
-                    module,
-                    store.clone(),
-                    &if_stmt.true_side,
-                    arena,
-                    collector,
-                )?;
-                let false_cond = arena.alloc(SLTNode::Unary(UnaryOp::LogicNot, cond_node));
-                let false_guard = if let Some(active) = saved_guard {
-                    arena.alloc(SLTNode::Binary(active, BinaryOp::LogicAnd, false_cond))
-                } else {
-                    false_cond
-                };
-                let mut false_sources = saved_guard_sources.clone();
-                false_sources.extend(sources.iter().copied());
-                collector.active_guard = Some(false_guard);
-                collector.active_guard_sources = false_sources;
-                let else_store = collect_comb_effects_statements(
-                    module,
-                    store,
-                    &if_stmt.false_side,
-                    arena,
-                    collector,
-                )?;
-                collector.active_guard = saved_guard;
-                collector.active_guard_sources = saved_guard_sources;
-                store = merge_symbolic_stores(
-                    module,
-                    &side_store,
-                    &else_store,
-                    bool_node(arena, true),
-                    &HashSet::default(),
-                    arena,
-                )?;
-            }
-            Statement::IfReset(_)
-            | Statement::TbMethodCall(_)
-            | Statement::Break
-            | Statement::Unsupported(_)
-            | Statement::Null => {}
-        }
-    }
-    Ok(store)
-}
-
-fn collect_comb_effects_for(
-    module: &Module,
-    mut store: SymbolicStore<VarId>,
-    for_stmt: &ForStatement,
-    arena: &mut SLTNodeArena<VarId>,
-    collector: &mut CombEffectCollector,
-) -> Result<SymbolicStore<VarId>, ParserError> {
-    let loop_width = for_stmt.var_type.total_width().unwrap_or(32);
-    let original_store = store.clone();
-    let ForRange::Forward {
-        start,
-        end,
-        inclusive,
-        ..
-    } = &for_stmt.range
-    else {
-        let (loop_effects, observer_start) =
-            collect_dynamic_for_effects(module, &store, for_stmt, arena, collector)?;
-        let (store, _, runner) = eval_for_with_effects(
-            module,
-            store,
-            BoundaryMap::default(),
-            for_stmt,
-            arena,
-            &loop_effects,
-        )?;
-        attach_loop_runner_to_first_observer(collector, observer_start, runner);
-        return Ok(store);
-    };
-    let Some(start) = const_for_bound_i64(start) else {
-        let (loop_effects, observer_start) =
-            collect_dynamic_for_effects(module, &store, for_stmt, arena, collector)?;
-        let (store, _, runner) = eval_for_with_effects(
-            module,
-            store,
-            BoundaryMap::default(),
-            for_stmt,
-            arena,
-            &loop_effects,
-        )?;
-        attach_loop_runner_to_first_observer(collector, observer_start, runner);
-        return Ok(store);
-    };
-    let Some(end) = const_for_bound_i64(end) else {
-        let (loop_effects, observer_start) =
-            collect_dynamic_for_effects(module, &store, for_stmt, arena, collector)?;
-        let (store, _, runner) = eval_for_with_effects(
-            module,
-            store,
-            BoundaryMap::default(),
-            for_stmt,
-            arena,
-            &loop_effects,
-        )?;
-        attach_loop_runner_to_first_observer(collector, observer_start, runner);
-        return Ok(store);
-    };
-    let final_end = if *inclusive { end + 1 } else { end };
-    for i in start..final_end {
-        let mut iter_store = store.clone();
-        let node = arena.alloc(SLTNode::Constant(
-            BigUint::from(i as u64),
-            BigUint::from(0u8),
-            loop_width,
-            for_stmt.var_type.signed,
-        ));
-        iter_store.insert(
-            for_stmt.var_id,
-            RangeStore::new(Some((node, HashSet::default())), loop_width),
-        );
-        store =
-            collect_comb_effects_statements(module, iter_store, &for_stmt.body, arena, collector)?;
-        store.remove(&for_stmt.var_id);
-    }
-    eval_for(
-        module,
-        original_store,
-        BoundaryMap::default(),
-        for_stmt,
-        arena,
-    )
-    .map(|(s, _)| s)
-}
-
-fn collect_dynamic_for_effects(
-    module: &Module,
-    store: &SymbolicStore<VarId>,
-    for_stmt: &ForStatement,
-    arena: &mut SLTNodeArena<VarId>,
-    collector: &mut CombEffectCollector,
-) -> Result<(Vec<SLTForEffect>, usize), ParserError> {
-    let Some(loop_width) = for_stmt.var_type.total_width() else {
-        return Ok((Vec::new(), collector.observers.len()));
-    };
-    let mut iter_store = store.clone();
-    let mut written_accesses = HashMap::default();
-    collect_written_accesses(module, &for_stmt.body, &mut written_accesses)?;
-    for (id, accesses) in written_accesses {
-        let width = resolve_total_width(module, &module.variables[&id])?;
-        let mut loop_store = RangeStore::new(None, width);
-        let mut covered = vec![false; width];
-        for access in accesses {
-            for slot in covered.iter_mut().take(access.msb + 1).skip(access.lsb) {
-                *slot = true;
-            }
-        }
-        let original = store
-            .get(&id)
-            .cloned()
-            .unwrap_or_else(|| RangeStore::new(None, width));
-        let mut bit = 0usize;
-        while bit < width {
-            if covered[bit] {
-                bit += 1;
-                continue;
-            }
-            let start = bit;
-            while bit < width && !covered[bit] {
-                bit += 1;
-            }
-            let end = bit - 1;
-            let access = BitAccess::new(start, end);
-            let parts = original.get_parts(access);
-            let (expr, sources) = combine_parts_with_default(id, access.lsb, parts, arena);
-            loop_store.update(access, Some((expr, sources)));
-        }
-        iter_store.insert(id, loop_store);
-    }
-    iter_store.insert(for_stmt.var_id, RangeStore::new(None, loop_width));
-    let observer_start = collector.observers.len();
-    let saved = collector.loop_effects.take();
-    collector.loop_effects = Some(Vec::new());
-    let _ = collect_comb_effects_statements(module, iter_store, &for_stmt.body, arena, collector)?;
-    let effects = collector.loop_effects.take().unwrap_or_default();
-    collector.loop_effects = saved;
-    Ok((effects, observer_start))
-}
-
-fn attach_loop_runner_to_first_observer(
-    collector: &mut CombEffectCollector,
-    observer_start: usize,
-    runner: Option<NodeId>,
-) {
-    let Some(runner) = runner else {
-        return;
-    };
-    if let Some(observer) = collector.observers[observer_start..]
-        .iter_mut()
-        .find(|observer| observer.captured_in_loop)
-    {
-        observer.loop_runner = Some(runner);
-    }
 }
 
 fn const_for_bound_i64(bound: &ForBound) -> Option<i64> {
